follows the advice of -Wsign-compare
Bug: 181177782
Test: None
Merged-In: Ie30386ce4d7a379f0e73ad8aed5a87348310ab81

+++ common/profiler/profiler.cc
@@ -194,7 +194,7 @@ void ProfilerImpl::End(const std::string name, int request_id) {
 
   std::lock_guard<std::mutex> lk(lock_);
 
-  if (index < timing_map_[name].size()) {
+  if (static_cast<std::size_t>(index) < timing_map_[name].size()) {
     TimeSlot& slot = timing_map_[name][index];
     slot.end += CurrentTime();
     slot.count++;
